replace double with int, version 1.1

+++ workspace-gateway/src/main/java/com/aliware/tianchi/util/ScalableSemaphore.java
@@ -3,15 +3,15 @@ package com.aliware.tianchi.util;
 import java.util.concurrent.Semaphore;
 
 public class ScalableSemaphore extends Semaphore {
-    public ScalableSemaphore(double permits) {
-        super((int)permits);
+    public ScalableSemaphore(int permits) {
+        super(permits);
     }
 
-    public void reducePermitsInternal(double num) {
-        super.reducePermits((int)num);
+    public void reducePermitsInternal(int num) {
+        super.reducePermits(num);
     }
 
-    public void increasePermits(double num) {
-        release((int)num);
+    public void increasePermits(int num) {
+        release(num);
     }
 }
